+++ src/api/integrations/chatbot/chatbot.router.ts
@@ -4,8 +4,8 @@ import { OpenaiRouter } from '@api/integrations/chatbot/openai/routes/openai.rou
 import { TypebotRouter } from '@api/integrations/chatbot/typebot/routes/typebot.router';
 import { Router } from 'express';
 
+import { EvolutionBotRouter } from './evolutionBot/routes/evolutionBot.router';
 import { FlowiseRouter } from './flowise/routes/flowise.router';
-import { GenericRouter } from './generic/routes/generic.router';
 
 export class ChatbotRouter {
   public readonly router: Router;
@@ -13,7 +13,7 @@ export class ChatbotRouter {
   constructor(...guards: any[]) {
     this.router = Router();
 
-    this.router.use('/generic', new GenericRouter(...guards).router);
+    this.router.use('/evolutionBot', new EvolutionBotRouter(...guards).router);
     this.router.use('/chatwoot', new ChatwootRouter(...guards).router);
     this.router.use('/typebot', new TypebotRouter(...guards).router);
     this.router.use('/openai', new OpenaiRouter(...guards).router);