+++ src/api/server.module.ts
@@ -19,10 +19,10 @@ import { ChatwootController } from './integrations/chatbot/chatwoot/controllers/
 import { ChatwootService } from './integrations/chatbot/chatwoot/services/chatwoot.service';
 import { DifyController } from './integrations/chatbot/dify/controllers/dify.controller';
 import { DifyService } from './integrations/chatbot/dify/services/dify.service';
+import { EvolutionBotController } from './integrations/chatbot/evolutionBot/controllers/evolutionBot.controller';
+import { EvolutionBotService } from './integrations/chatbot/evolutionBot/services/evolutionBot.service';
 import { FlowiseController } from './integrations/chatbot/flowise/controllers/flowise.controller';
 import { FlowiseService } from './integrations/chatbot/flowise/services/flowise.service';
-import { GenericController } from './integrations/chatbot/generic/controllers/generic.controller';
-import { GenericService } from './integrations/chatbot/generic/services/generic.service';
 import { OpenaiController } from './integrations/chatbot/openai/controllers/openai.controller';
 import { OpenaiService } from './integrations/chatbot/openai/services/openai.service';
 import { TypebotController } from './integrations/chatbot/typebot/controllers/typebot.controller';
@@ -116,8 +116,8 @@ export const openaiController = new OpenaiController(openaiService, prismaReposi
 const difyService = new DifyService(waMonitor, configService, prismaRepository);
 export const difyController = new DifyController(difyService, prismaRepository, waMonitor);
 
-const genericService = new GenericService(waMonitor, configService, prismaRepository);
-export const genericController = new GenericController(genericService, prismaRepository, waMonitor);
+const evolutionBotService = new EvolutionBotService(waMonitor, configService, prismaRepository);
+export const evolutionBotController = new EvolutionBotController(evolutionBotService, prismaRepository, waMonitor);
 
 const flowiseService = new FlowiseService(waMonitor, configService, prismaRepository);
 export const flowiseController = new FlowiseController(flowiseService, prismaRepository, waMonitor);