+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 * Improved layout manager
 * Translation in manager: English, Portuguese, Spanish and French
-* Generic Bot Integration
+* Evolution Bot Integration
 * Option to disable chatwoot bot contact with CHATWOOT_BOT_CONTACT
 * Added flowise integration
 * Added evolution channel on instance create